dt/requirements.txt
Django==2
gunicorn==19.6.0
djangorestframework
requests
beautifulsoup4